Rename listndirname.c to list n dirname.c
Write a C program that behaves like a shell which displays the command prompt ‘myshell$’. It accepts the command, tokenize the command line and execute it by creating the child process.
Also implementthe additional command ‘list’ as myshell$ list f dirname: It will display filenames in a given directory. myshell$ list n dirname: It will count the number of entries in a given
directory. myshell$ list i dirname: It will display filenames and theirinode number for the filesin
a given directory.